refactor: redesign the src dir structure
