--- src/_internal/index.ts
@@ -1,1 +1,0 @@
-export * from './resizable'
