--- src/_internal/resizable/border/config.ts
@@ -1,9 +1,0 @@
-import type { ResizableBorderConfig } from '../config'
-
-export const defaultConfig: ResizableBorderConfig = {
-  render: true,
-  style: {
-    headless: false,
-    size: 5,
-  },
-}
