+++ src/components/index.ts
@@ -1,3 +1,1 @@
-export {
-  Resizable,
-} from './resizable'
+export * from './resizable'
