+++ src/composables/index.ts
@@ -1,1 +1,2 @@
-export { useResizable } from '@/_internal'
+export * from './cursor'
+export * from './resizable'
