+++ src/types/index.ts
@@ -1,2 +1,3 @@
-export type { ResizableBorderConfig, ResizableConfig } from '../_internal/resizable/config'
-export type { ResizableEl } from '../_internal/resizable'
+export * from './config'
+export * from './edge'
+export * from './resizable'
