+++ src/types/resizable.ts
@@ -1,0 +1,3 @@
+export interface Position { x: number; y: number }
+
+export type ResizableEl = HTMLElement
