+++ src/utils/dom/edge.ts
@@ -1,22 +1,7 @@
-export enum BaseEdge {
-  LEFT = 'left',
-  RIGHT = 'right',
-  TOP = 'top',
-  BOTTOM = 'bottom',
-}
-
-export enum ExtendedEdge {
-  TOP_LEFT = `${BaseEdge.TOP}-${BaseEdge.LEFT}`,
-  TOP_RIGHT = `${BaseEdge.TOP}-${BaseEdge.RIGHT}`,
-  BOTTOM_LEFT = `${BaseEdge.BOTTOM}-${BaseEdge.LEFT}`,
-  BOTTOM_RIGHT = `${BaseEdge.BOTTOM}-${BaseEdge.RIGHT}`,
-}
-
-export type Edge = BaseEdge | ExtendedEdge
+import type { IsInEdgeResult } from '@/types'
+import { BaseEdge, ExtendedEdge } from '@/types'
 
 const TOLERANCE = 5
-
-export type IsInEdgeResult = Record<Edge, boolean>
 
 export function isInAround(el: HTMLElement, x: number, y: number) {
   const { left, right, top, bottom } = el.getBoundingClientRect()
