+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './dom'
 export * from './common'
+export * from './render'
+export * from './config'
